// src/types/personType.tsx
interface PersonType {
  id: number;
  name: string;
  email: string;
  cpf: string;
  birth_date: string;
  gender: string;
  cell_phone: string;
  state: string;
  city: string;
  neighborhood: string;
  street: string;
  number: string;
  zip_code: string;
  complement: string;
  reference_point: string;
  observations: string;
}

export type { PersonType };